# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Game API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "tags": [
    {
      "name": "Authentication",
      "description": "Auth endpoints"
    },
    {
      "name": "Games",
      "description": "Game-related endpoints"
    },
    {
      "name": "Game Rating",
      "description": "Game ratings"
    }
  ],
  "components": {
    "schemas": {},
    "parameters": {}
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string"
                    },
                    "playerId": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "sessionId",
                    "playerId",
                    "expiresAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "def": {
                "type": "object",
                "shape": {
                  "error": {
                    "def": {
                      "type": "string"
                    },
                    "type": "string",
                    "format": null,
                    "minLength": null,
                    "maxLength": null
                  },
                  "message": {
                    "def": {
                      "type": "optional",
                      "innerType": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      }
                    },
                    "type": "optional"
                  }
                }
              },
              "type": "object"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "content": {}
        },
        "responses": {
          "200": {
            "description": "session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string"
                    },
                    "playerId": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "sessionId",
                    "playerId",
                    "expiresAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/games/{gameId}/ratings": {
      "post": {
        "tags": [
          "Game Rating"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "gameId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rating": {
                    "type": "number"
                  },
                  "comment": {
                    "type": "string"
                  }
                },
                "required": [
                  "rating"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "game_id": {
                      "type": "string"
                    },
                    "player_id": {
                      "type": "string"
                    },
                    "rating": {
                      "type": "number"
                    },
                    "comment": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "created_at",
                    "game_id",
                    "player_id",
                    "rating"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Game Rating"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "gameId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "rating": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "rating"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/games": {
      "get": {
        "description": "list all games",
        "tags": [
          "Games"
        ],
        "responses": {
          "200": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "created_at": {
                        "type": [
                          "string",
                          "null"
                        ],
                        "format": "date-time"
                      },
                      "average_rating": {
                        "type": "number"
                      },
                      "rating_count": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "id",
                      "title",
                      "description",
                      "created_at",
                      "average_rating",
                      "rating_count"
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "def": {
                "type": "object",
                "shape": {
                  "error": {
                    "def": {
                      "type": "string"
                    },
                    "type": "string",
                    "format": null,
                    "minLength": null,
                    "maxLength": null
                  },
                  "message": {
                    "def": {
                      "type": "optional",
                      "innerType": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      }
                    },
                    "type": "optional"
                  }
                }
              },
              "type": "object"
            }
          }
        }
      }
    },
    "/games/{gameId}": {
      "get": {
        "description": "get details for a game",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "gameId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    },
                    "average_rating": {
                      "type": "number"
                    },
                    "rating_count": {
                      "type": "number"
                    },
                    "game_ratings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "player_id": {
                            "type": "string"
                          },
                          "rating": {
                            "type": "number"
                          },
                          "comment": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "created_at": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "date-time"
                          },
                          "player": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "username": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "username"
                            ]
                          }
                        },
                        "required": [
                          "player_id",
                          "rating",
                          "comment",
                          "created_at",
                          "player"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "description",
                    "created_at",
                    "average_rating",
                    "rating_count",
                    "game_ratings"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/games/{gameId}/scenarios": {
      "get": {
        "description": "get all scenarios for a game",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "gameId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "create a new scenario (requires admin role)",
        "security": [
          {
            "cookieAuth": [
              "ADMIN"
            ]
          }
        ],
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "gameId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    },
    "/games/{gameId}/scenarios/{scenarioId}/start-new": {
      "post": {
        "description": "start playing a scenario",
        "tags": [
          "Games"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "scenarioId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "gameId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "no description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            }
          }
        }
      }
    }
  },
  "webhooks": {}
}